+++ packages/familie-header/src/header/index.ts
@@ -1,0 +1,1 @@
+export * from './Header';
