+++ packages/familie-header/src/icons/Ikon.tsx
@@ -9,12 +9,10 @@ export interface IkonProps {
     children?: ReactNode | ReactNode[];
 }
 
-const Ikon = ({ children, width = 16, height = 16, viewBox = 24, className='' }: IkonProps) => {
+export const Ikon = ({ children, width = 16, height = 16, viewBox = 24, className='' }: IkonProps) => {
     return (
         <svg xmlns="http://www.w3.org/2000/svg" width={width} height={height} viewBox={`0 0 ${viewBox} ${viewBox}`} className= {className}>
             {children}
         </svg>
     );
 };
-
-export default Ikon;
