+++ packages/familie-header/src/icons/IkonSøk.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
-import Ikon, { IkonProps } from './Ikon';
+import {Ikon, IkonProps } from './Ikon';
 
 const navGra60= '#78706a';
 
-const IkonSøk = ({ color = navGra60, width = 16, height = 16 }: IkonProps) => {
+export const IkonSøk = ({ color = navGra60, width = 16, height = 16 }: IkonProps) => {
     return (
         <Ikon width={width} height={height}>
             <path
@@ -13,5 +13,3 @@ const IkonSøk = ({ color = navGra60, width = 16, height = 16 }: IkonProps) => {
         </Ikon>
     );
 };
-
-export default IkonSøk;
