+++ packages/familie-header/src/icons/index.ts
@@ -1,0 +1,6 @@
+export * from './Ikon';
+export * from './IkonFeil';
+export * from './IkonGyldig';
+export * from './IkonSpinner';
+export * from './IkonSystem';
+export * from './IkonSøk';
