+++ packages/familie-header/src/index.ts
@@ -1,7 +1,3 @@
-import Header from './header/Header';
-import Infokort from './søk/Infokort';
-import Infostripe, { InfoType } from './søk/Infostripe';
-import Søk from './søk/Søk';
-
-export { Søk, Infokort, Infostripe, InfoType};
-export default Header;
+export * from './header';
+export * from './icons';
+export * from './søk';
