+++ packages/familie-header/src/søk/Søk.less
@@ -12,6 +12,13 @@
     left: 0;
     position: absolute;
     top: 50%;
+    transform: translateY(-50%);
+  }
+
+  &__spinner{
+    left: 0;
+    position: absolute;
+    top: 10%;
     transform: scale(0.8, 0.8) translateY(-50%);
   }
