+++ packages/familie-header/src/søk/index.ts
@@ -1,0 +1,3 @@
+export * from './Infokort';
+export * from './Infostripe';
+export * from './Søk';
